requirements.txt
django<2.3
djangorestframework<3.11
redis<3.4

aiohttp<3.7
aioredis<1.4

srvlookup<2.1
